fix switch and 3ds for reals yo
