+++ platform/switch/source/modules/graphics.cpp
@@ -803,15 +803,14 @@ int Graphics::GetRendererInfo(lua_State * L)
 
 void Graphics::Exit()
 {
+    Display::Exit();
+
     TTF_Quit();
     SDL_Quit(); //kill remaining stuff
 }
 
 int Graphics::Register(lua_State * L)
 {
-    if (SDL_Init(SDL_INIT_VIDEO) != 0)
-        luaL_error(L, "Failed to initialize SDL2!");
-
     TTF_Init();
 
     stack.reserve(16);
