+++ source/main.cpp
@@ -51,7 +51,7 @@ int main(int argc, char * argv[])
     love_preload(L, LuaSocket::InitSocket, "socket");
     love_preload(L, LuaSocket::InitHTTP,   "socket.http");
 
-    Logger::Initialize();
+    //Logger::Initialize();
 
     char * path = (argc == 2) ? argv[1] : argv[0];
     Filesystem::Initialize(path);
@@ -73,7 +73,7 @@ int main(int argc, char * argv[])
 
     Love::Exit(L);
 
-    Logger::Exit();
+    //Logger::Exit();
 
     return 0;
 }
