+++ source/modules/love.cpp
@@ -11,20 +11,10 @@
 #include "modules/joystick.h"
 #include "modules/timer.h"
 #include "modules/window.h"
-#include "modules/display.h"
 
 #include "socket/luasocket.h"
 
 #include "nogame_lua.h"
-
-void Love::InitModules(lua_State * L)
-{
-    for (int i = 0; modules[i].name; i++)
-    {
-        if (modules[i].init)
-            modules[i].init();
-    }
-}
 
 void Love::InitConstants(lua_State * L)
 {
@@ -86,8 +76,6 @@ int Love::Initialize(lua_State * L)
         strcat(modname, modules[i].name);
         love_preload(L, modules[i].reg, modname);
     }
-
-    Display::Initialize();
 
     LuaSocket::Initialize();
 
@@ -172,6 +160,5 @@ void Love::Exit(lua_State * L)
 
     lua_close(L);
 
-    Display::Exit();
     Graphics::Exit();
 }
